+++ plan.md
@@ -1,1 +1,1 @@
-`src/shared/ui/ProfileHeader` を `src/pages/profile/ui` へ移動する
+`src/shared/ui/` の中から特定のページでしか使われていないコンポーネントを `src/pages/〇〇/ui` へ移動する。
