+++ src/pages/follows/index.ts
@@ -1,1 +1,2 @@
 export { FollowsPageClient } from "./ui/FollowsPageClient";
+export { Follows } from "./ui/Follows";
