+++ src/pages/follows/ui/Follows.tsx
@@ -1,11 +1,12 @@
+"use client";
+
 import { Box, Tabs } from "@mantine/core";
 import useSWR from "swr";
 
 import type { User } from "@/shared/entities";
 import { fetcher } from "@/shared/lib";
 import { baseURL } from "@/shared/lib/const";
-import { CenteredLoader } from "@/shared/ui/CenteredLoader/CenteredLoader";
-import { UsersList } from "@/shared/ui/UsersList/UsersList";
+import { CenteredLoader, UsersList } from "@/shared/ui";
 
 interface FollowsProps {
 	currentUserId: string;
