+++ src/pages/posts/index.ts
@@ -1,1 +1,4 @@
 export { PostDetailPageClient } from "./ui/PostDetailPageClient";
+export { CommentForm } from "./ui/CommentForm";
+export { CommentList } from "./ui/CommentList";
+export { PreviousPageHeader } from "./ui/PreviousPageHeader";
