+++ src/pages/profile/index.ts
@@ -1,3 +1,5 @@
 export { ProfilePageClient } from "./ui/ProfilePageClient";
 export { Profile } from "./ui/Profile";
 export { ProfileHeader } from "./ui/ProfileHeader";
+export { ProfileActionsButton } from "./ui/ProfileActionsButton";
+export { ProfileStats } from "./ui/ProfileStats";
